fixed model relationship name

+++ app/Http/Controllers/ProfileController.php
@@ -11,22 +11,32 @@ use Illuminate\Support\Facades\Storage;
 
 class ProfileController extends Controller
 {
+    /**
+     * @param $id
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function index($id){
         $user = User::find($id);
         $post_by_user = Post::where("user_id", $id)->when(request()->search, function($query){
             $search_key = request()->search;
             return $query->where("title", "LIKE", "%$search_key%")->orWhere("description", "LIKE", "%$search_key%");
-        })->get();
+        })->latest()->get();
         return view('profile.index', compact("user", "post_by_user"));
     }
 
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function profilePictureUpdate(Request $request){
         $request->validate([
-            "profile_picture" => "required|mimetypes:image/jpeg,image/png|file|max:2500|dimensions:ratio=1/1"
+            "profile_picture" => "required|mimetypes:image/jpeg,image/png|file|max:2500"
         ]);
         $dir = "public/profile-picture/";
 
-        Storage::delete($dir . Auth::user()->profile_picture);
+        if (!Auth::user()->profile_picture == "default-profile.jpg") {
+            Storage::delete($dir . Auth::user()->profile_picture);
+        }
 
         $newName = uniqid() . "_profile-picture." . $request->file("profile_picture")->getClientOriginalExtension();
         $request->file("profile_picture")->storeAs($dir, $newName);
@@ -38,7 +48,10 @@ class ProfileController extends Controller
         return redirect()->route("profile.edit");
     }
 
-
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function profileUpdate(Request $request){
         $request->validate([
             "name" => "required|string|min:1|max:100",
